pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
pandas = "^1.5.3"
psycopg2 = "^2.9.5"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.1"

[tool.poetry.dev-dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
